+++ app/src/main/java/com/github/marmaladesky/data/RevelationData.java
@@ -1,8 +1,11 @@
 package com.github.marmaladesky.data;
 
 import android.content.ContentResolver;
+import android.content.Context;
 import android.net.Uri;
+
 import com.github.marmaladesky.Cryptographer;
+
 import org.simpleframework.xml.*;
 import org.simpleframework.xml.core.Persister;
 
@@ -10,26 +13,27 @@ import java.io.*;
 import java.util.List;
 import java.util.UUID;
 
+import de.igloffstein.maik.aRevelation.Helper.ARevelationHelper;
 import lombok.Getter;
 
-@Root(name="revelationdata")
+@Root(name = "revelationdata")
 @Order(attributes = {"version", "dataversion"})
 public class RevelationData implements Serializable {
 
     private final String uuid = UUID.randomUUID().toString();
 
-    @Attribute(name="version")
+    @Attribute(name = "version")
     private String version;
 
-    @Attribute(name="dataversion")
+    @Attribute(name = "dataversion")
     private String dataversion;
 
     @Getter
     @ElementList(inline = true)
     private List<Entry> entries;
 
-    public RevelationData(@Attribute(name="version") String version,
-                          @Attribute(name="dataversion") String dataversion,
+    public RevelationData(@Attribute(name = "version") String version,
+                          @Attribute(name = "dataversion") String dataversion,
                           @ElementList(inline = true) List<Entry> entries) {
         this.version = version;
         this.dataversion = dataversion;
@@ -40,19 +44,19 @@ public class RevelationData implements Serializable {
         return getEntryById(entries, uuid);
     }
 
-    public void addEntry(Entry entry){
+    public void addEntry(Entry entry) {
         entries.add(entry);
     }
 
     private static Entry getEntryById(List<Entry> list, String uuid) {
-        if(list != null)
-            for(Entry e : list) {
-                if(e.type.equals(Entry.TYPE_FOLDER)) {
+        if (list != null)
+            for (Entry e : list) {
+                if (e.type.equals(Entry.TYPE_FOLDER)) {
                     Entry n = getEntryById(e.list, uuid);
-                    if(n != null)
+                    if (n != null)
                         return n;
                 }
-                if(e.getUuid().equals(uuid))
+                if (e.getUuid().equals(uuid))
                     return e;
             }
         return null;
@@ -60,19 +64,19 @@ public class RevelationData implements Serializable {
 
     public FieldWrapper getFieldById(String uuid) throws Exception {
         FieldWrapper fw = getFieldById(uuid, entries);
-        if(fw != null)
+        if (fw != null)
             return fw;
         else
             throw new Exception("Cannot find field with id=" + uuid);
     }
 
     private static FieldWrapper getFieldById(String uuid, List<Entry> entries) {
-        for(Entry e : entries) {
-            if(e.list != null) {
+        for (Entry e : entries) {
+            if (e.list != null) {
                 FieldWrapper fw = getFieldById(uuid, e.list);
-                if(fw != null) return fw;
+                if (fw != null) return fw;
             }
-            if(e.fields != null) {
+            if (e.fields != null) {
                 for (Field f : e.fields) {
                     if (e.getUuidName().equals(uuid)) {
                         return new FieldWrapper(Entry.PROPERTY_NAME, e);
@@ -97,11 +101,11 @@ public class RevelationData implements Serializable {
     }
 
     public List<Entry> getEntryGroupById(String uuid) throws Exception {
-        if(this.uuid.equals(uuid)) {
+        if (this.uuid.equals(uuid)) {
             return entries;
         } else {
             List<Entry> l = getEntryGroupById(entries, uuid);
-            if(l != null)
+            if (l != null)
                 return getEntryGroupById(entries, uuid);
             else
                 throw new Exception("Cannot find group with id = " + uuid);
@@ -109,9 +113,9 @@ public class RevelationData implements Serializable {
     }
 
     private static List<Entry> getEntryGroupById(List<Entry> entries, String uuid) {
-        for(Entry e : entries) {
-            if(e.type.equals(Entry.TYPE_FOLDER)) {
-                if(e.getUuid().equals(uuid)) {
+        for (Entry e : entries) {
+            if (e.type.equals(Entry.TYPE_FOLDER)) {
+                if (e.getUuid().equals(uuid)) {
                     return e.list;
                 } else {
                     List<Entry> l = getEntryGroupById(e.list, uuid);
@@ -127,23 +131,23 @@ public class RevelationData implements Serializable {
     }
 
     public boolean isEdited() {
-        for(Entry e : entries) if(e.isEdited()) return true;
+        for (Entry e : entries) if (e.isEdited()) return true;
         return false;
     }
 
-    public void save(String file, String password, ContentResolver contentResolver) throws Exception {
+    public void save(Context context, String file, String password) throws Exception {
         Serializer serializer = new Persister();
         Writer writer = new StringWriter();
         serializer.write(this, writer);
         byte[] encrypted = Cryptographer.encrypt(writer.toString(), password);
 
-        OutputStream fop = contentResolver.openOutputStream(Uri.parse(file));
+        OutputStream fop = context.getContentResolver().openOutputStream(Uri.parse(file));
         fop.write(encrypted);
         fop.close();
         cleanUpdateStatus();
     }
 
     private void cleanUpdateStatus() {
-        for(Entry e : entries) e.cleanUpdateStatus();
+        for (Entry e : entries) e.cleanUpdateStatus();
     }
 }